test: migrate integrations-telegram-delete to bdd
